// types.ts
export interface Branch {
  id: number;
  projectId: number;
  name: string;
}
// ...
export interface SupplyChainItem {
  id: number
  sku?: string
  gtin?: string
  batchNumber?: string
  serialNumber?: string
  productName: string
  quantity: number
  unit?: string
  manufacturer?: string
  originCountry?: string
  manufactureDate?: string
  expiryDate?: string
  currentStatus?: string
  transportMode?: string
  created_at?: string
  updated_at?: string
}